requirements.txt
crewai
crewai-tools
langchain-google-genai
google-generativeai
gradio
elevenlabs
python-dotenv
firecrawl